requirements.txt
flask==3.1.0
google-cloud-dialogflow==2.41.1